refactor: eliminar el caso de uso y DTO de recuperación de contraseña, y ajustar el módulo de usuarios para reflejar estos cambios.
