--- users/src/users/dto/forgot-password.dto.ts
@@ -1,7 +1,0 @@
-import { IsEmail, IsNotEmpty } from 'class-validator';
-
-export class ForgotPasswordDto {
-  @IsEmail({}, { message: 'email is not valid' })
-  @IsNotEmpty({ message: 'email is required' })
-  email: string;
-}
