+++ users/src/users/users.module.ts
@@ -1,5 +1,6 @@
 import { Module } from '@nestjs/common';
 import { NatsModule } from 'src/transports/nats.module';
+import { ForgotPasswordUseCase } from '../auth/service/use-cases/forgot-password.use-cases';
 import { MockEmailService } from '../common/services/mock-email.service';
 import { NodemailerService } from '../common/services/nodemailer.service';
 import { UserBaseService } from '../common/services/user-base.service';
@@ -8,7 +9,6 @@ import { SharedModule } from '../shared/shared.module';
 import { UsersController } from './controller/users.controller';
 import { UserRepository } from './repository/users.repository';
 import { CreateUserUseCase } from './service/use-cases/create-user.use-cases';
-import { ForgotPasswordUseCase } from './service/use-cases/forgot-password.use-cases';
 import { PingUseCase } from './service/use-cases/ping';
 import { ResendVerificationUseCase } from './service/use-cases/resend-verification.use-cases';
 import { VerifyUserUseCase } from './service/use-cases/verify-user.use-cases';
@@ -31,6 +31,6 @@ import { UsersService } from './service/users.service';
     },
   ],
   imports: [NatsModule, SharedModule, ProfileModule],
-  exports: [UserRepository],
+  exports: [UserRepository, UserBaseService],
 })
 export class UsersModule {}
